Updated `module/xplm:run()'
It is now possible to call executables within the given XPLM
environment, rather than assuming that the argument to run is
definitely the path to a script.

If the given argument is not a path to a script (i.e., it does not
exist), then the run function will just attempt to call it and assume
that it exists in the path.

This behaviour can still be improved, but good enough for now.