Out with `util:json' and `jsontool', in with `jq'
We will now simply depend on jq being available, and jq can be
called directly.  Before, the utility used to provide commandline json
capabilities was not good enough to hardcode, and thus it was
encapsulated in the this helper function, however `jq' is of very good
quality, so no need for the extra layer of abstraction.