Handle the sourcing of bad modules
Modules (specially user modules) can be modified and become `bad',
i.e., exit and break the code execution path defined in simbol.

To handle this, we now source the module in a subshell first, if that
goes well, then we source it again in the parent.  If not, we log the
errors (if any).