Fixed `:hgd:save'; now a little more intuitive
The issue were a formula that resoved to a zero set resulted in the
same `session' of the same name not being deleted has been resolved.

Also, `:hgd:save' now calls `:hgd:resolve' on the *new* formula, so
the user doesn't have to (annoyingly) run `:hgd:refresh' manually
each time the save a new `session'.

Finally, we now return success where the new formula resolves to a
non-zero set.