Implemented the validation function
We now not only test for the minimum bash version, but assert that
certain operations execute with expected side-effect.  The set of
tests here is only a starting point, over time more will be added.

This is almost like a pre-execution sanity check, to confirm that your
version of bash will not run into trouble and result in unexpected
behaviour.