Use `env' to determine which `bash' to run with
This should reduce load/run-time if the preferred bash is not
`/bin/bash', but more importantly will allow for environment-exporting
of non-POSIX keys from the environement; namely those defined with
`declare -a' and `declare -A'.

That is, if `/bin/bash' is ancient, then going from a supported
version, to an unsupported version, and then back to a supported
version is both wasteful, and will garanteed to fail.

Work pertaining to this will be added in this branch.